move benchmarks.py to root level
